XKCD Alpha Webscraper
This one gets all of the information from the most recent xkcd comic off of the front page of explainxkcd. If we were to replace the "Main_Page" part of the url with any number from 1 to 1747, the code would fail. This only works to get out the most recent transcript and title.